ai-service/requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.0

# AI/ML - Core
openai==1.59.6
anthropic==0.42.0

# LangChain + LangGraph (Agentic RAG)
langchain==0.3.13
langchain-openai==0.2.14
langchain-anthropic==0.3.4
langchain-community==0.3.13
langgraph==0.2.60
langsmith==0.2.10

# Embeddings - Voyage (top quality)
voyageai==0.3.2

# Vector DB
pgvector==0.3.6
psycopg2-binary==2.9.10

# Database
sqlalchemy==2.0.37
asyncpg==0.30.0

# Research Tools
exa-py==1.9.0
firecrawl-py==1.9.0

# Search (fallback)
tavily-python==0.5.0

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.4.0

# MCP (Model Context Protocol)
mcp[cli]>=1.9.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
tiktoken==0.8.0
tenacity==9.0.0
